Map not showing bug (#1263)
* Better map management

PlaceCal uses turbo which breaks the leaflet map component on events and
partner pages. New JS has been written that uses the same map object on
multiple pages. This way the page can be loaded by turbo and the map
will 're-initialize' on each page load.

* Map JS loading and setting up on page loads

When the page is initially:
if there is a map then it is configured and shown
if there is no map then the setup is skipped

When navigation happens (non page load, through turbo)
if there is a map then it is activated and moved to the right place
if there is no map (and the map was active) then the map is hidden

tries to clear any markers that may already exist on the map

* Map styling in application assets now

* Removed mountain view map component

* Fixed tests, tidied stylesheets

* Mapbox API token is now pulled from ENV

Be sure to update your environment or it will complain

Also silences map API in test mode