Demonstration of system tests.
You will need chrome installed.

To run use command
```
  rails test:system test/system/create_admin_users_test.rb
```

These tests do not run when just `rails test` is invoked.
And yes they take a long time to start up and that can cause a timeout
(on my system). So if the tests fail on the first try then try again and
hopefully the browser will be cached the second time.