Fix: set up domain and build partner logo url (#1465)
Fix: Generate URLs using the correct domain

Fixes environment configurations so the Rails application domain is correct for the environment (dev, test, prod and staging). Application domain is then used for URL generation in app and email links.

Use the full URL when generating links to resources in API requests.