Fix: Address lookup failure recovery (#1495)
* Fix: Address lookup failure recovery

If an address is created that does not map to a neighbourhood in
our database, just return nil to the callee.

In practice this is mainly used by the importer which does not
deal well with ActiveRecord exceptions. This is by design as any
AR problems should not be happening in the importer!

So now any failed lookups (by the importer) will cause it to mark
the calendar as failed and move on.