Removed broken test
Something in the component controller test was failing. Looks like
it was to do with how rails autoloads classes based on name.
We are not really using this functionality and to expidite this
branch merge I just took it out.
Really we should take all of this component styleguide stuff out
as we don't use it.