Warning admins about clearing partner addresses
If the admin is not root or does not own a partner then check to see if
the admin can "see" the partner via service areas. If they cannot then
warn the user that clearing the address will stop the admin from being
able to access this partner at all.