Re-add ERB component file, Rubocop and test fixes
Re-added the ERB component file so that there's a clear
separation of concerns. This also allows us to keep the code
easy to read; in case the template grows bigger, it will still
be readable.

Fixed the Rails/OutputSafety cop by moving the sanitize call
to within the template instead of trying to do it with the component

Also switched to using the `address` factory instead of a string
value; the value the component expects should be an instance of
the Address model and consequently the `all_address_lines` method
can be called on it.