Move template up a level, add all arguments
Thanks to Rails auto-loading, it seems the AdminIndexComponent
takes in more arguments than it appeared at first glance.
As a result, in order to safely migrate the behaviour,
we've added *all* the arguments that can be taken so that
the controller tests pass.

At a future date, we can remove the disabled cop and refactor
the UI accordingly.

Also moved the admin_index_component up a level so that
it can be found as well as updating the component calls.