chore: Migrate AdminIndexComponent to use ViewComponent (#2666)
* Migrate AdminIndexComponent to use ViewComponent

Migrated the AdminIndexComponent to use ViewComponent
instead of MountainView.

* Remove unneeded files

Now that we have moved over to ViewComponent, we no longer
need the old directory from mountain_view

* Move template up a level, add all arguments

Thanks to Rails auto-loading, it seems the AdminIndexComponent
takes in more arguments than it appeared at first glance.
As a result, in order to safely migrate the behaviour,
we've added *all* the arguments that can be taken so that
the controller tests pass.

At a future date, we can remove the disabled cop and refactor
the UI accordingly.

Also moved the admin_index_component up a level so that
it can be found as well as updating the component calls.

* Add test, change model_name method

Added the component test and changed the model_name
method as `.chop` cuts a letter off the model name
and `to_s` returns the in-memory address of the Collection
instance.

* Go back to using to_s

Went back to using to_s because that breaks existing
tests that rely on its behaviour.

* Remove test as it is more trouble than it is worth

Removing the presence test as it is more trouble to get it
to run than the value it would provide.

Should probably be fixed in a subsequent epic to make sure
that all components are covered by tests.

---------

Co-authored-by: Dr Kim Foale <kim@gfsc.studio>